This is a gitignore
Gitignore is a way to ignore files in your git repo.
I had some "artifacts" from my Editor that are of no importance to the code. They would have been in the `.idea` directory. I "gitignored" that directory in order to say, "Hey, these files are only important to me, and not Ellen"